+++ src/time.ts
@@ -1,36 +1,75 @@
 import { calc } from './number';
 
 /**
+ * 将字符串或数字解析成日期
+ * @param str 输入的日期字符串，如'2014-09-13'，也可以是一个时间戳
+ * @param fmt 字符串格式，支持如下：y、M、d、H、m、s、S，不支持w和q，不传时会根据str的格式尽量自动判断
+ * @returns 解析后的Date类型日期
+ */
+export const parseDate = (str: Date | string | number, fmt?: string): Date | null => {
+	if (!str) {
+		return null;
+	}
+	// 兼容已经是日期类型，直接返回
+	if (str instanceof Date) {
+		return str;
+	}
+	// 如果fmt为空，则根据str的格式尽量自动判断
+	if (!fmt) {
+		let date: Date = null;
+		// 如果是数字格式的时间戳
+		if (typeof str === 'number') {
+			// 1687682453445
+			date = new Date(str);
+		} else if (typeof str === 'string') {
+			// 如果是字符串格式的时间戳
+			if (/^\d{12,13}$/g.test(str)) {
+				// '1687682453445'
+				date = new Date(parseInt(str, 10));
+			} else if (/^.{10}T.{8,12}Z?$/g.test(str)) {
+				// '2019-01-01T00:00:00.000Z'
+				date = new Date(str);
+			} else if (/^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$/g.test(str)) {
+				// 2025-08-12 12:23:44
+				fmt = 'yyyy-MM-dd HH:mm:ss';
+			}
+		}
+		if (!fmt) {
+			// 如果日期合法
+			if (date instanceof Date && !isNaN(date?.getTime())) {
+				return date;
+			}
+			console.error('Auto parseDate error: invalid date:', str);
+			return null;
+		}
+	}
+	// 如果fmt不为空，则严格根据fmt的格式解析
+	const obj: any = { y: 0, M: 1, d: 0, H: 0, h: 0, m: 0, s: 0, S: 0 };
+	fmt.replace(/([^yMdHmsS]*?)(([yMdHmsS])\3*)([^yMdHmsS]*?)/g, (m, $1, $2, $3, $4) => {
+		str = `${str || ''}`.replace(new RegExp(`${$1}(\\d{${$2.length}})${$4}`), (_m, _$1) => {
+			obj[$3] = parseInt(_$1, 10);
+			return '';
+		});
+		return '';
+	});
+	obj.M--; // 月份是从0开始的，所以要减去1
+	const date = new Date(obj.y, obj.M, obj.d, obj.H, obj.m, obj.s);
+	if (obj.S !== 0) {
+		date.setMilliseconds(obj.S); // 如果设置了毫秒
+	}
+	return date;
+};
+
+/**
  * 将日期格式化成指定格式的字符串
  * @param date 要格式化的日期，不传时默认当前时间，也可以是一个时间戳等
  * @param fmt 目标字符串格式，支持的字符有：y,M,d,q,w,H,h,m,S，默认：yyyy-MM-dd HH:mm:ss
  * @returns 返回格式化后的日期字符串
  */
 export const formatDate = (date?: Date | number | string, fmt?: string) => {
+	date = parseDate(date);
 	if (!date) {
 		return '';
-	}
-	if (typeof date === 'number') {
-		// 1687682453445
-		date = new Date(date);
-	} else if (typeof date === 'string') {
-		if (/^\d{12,13}$/g.test(date)) {
-			// '1687682453445'
-			date = new Date(parseInt(date, 10));
-		} else if (/^.{10}T.{8,12}Z?$/g.test(date) || /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$/g.test(date)) {
-			// '2019-01-01T00:00:00.000Z'
-			// 2025-08-12 12:23:44
-			date = new Date(date);
-		} else {
-			// 其它不能处理的字符串类型，直接原样返回
-			return date;
-		}
-	}
-	if (!(date instanceof Date)) {
-		throw new Error('formatDate error: not date.');
-	}
-	if (isNaN(date?.getFullYear())) {
-		throw new Error('formatDate error: invalid date.');
 	}
 	fmt = fmt || 'yyyy-MM-dd HH:mm:ss';
 	const obj: any = {
@@ -61,29 +100,32 @@ export const formatDate = (date?: Date | number | string, fmt?: string) => {
 	return fmt;
 };
 
+
 /**
- * 将字符串解析成日期
- * @param str 输入的日期字符串，如'2014-09-13'
- * @param fmt 字符串格式，默认'yyyy-MM-dd'，支持如下：y、M、d、H、m、s、S，不支持w和q
- * @returns 解析后的Date类型日期
+ * 将一个日期格式化成友好格式，比如，1分钟以内的返回“刚刚”，
+ * 当天的返回时分，当年的返回月日，否则，返回年月日
+ * @param {Object} date
  */
-export const parseDate = (str: string, fmt?: string) => {
-	fmt = fmt || 'yyyy-MM-dd';
-	const obj: any = { y: 0, M: 1, d: 0, H: 0, h: 0, m: 0, s: 0, S: 0 };
-	fmt.replace(/([^yMdHmsS]*?)(([yMdHmsS])\3*)([^yMdHmsS]*?)/g, (m, $1, $2, $3, $4) => {
-		str = str.replace(new RegExp(`${$1}(\\d{${$2.length}})${$4}`), (_m, _$1) => {
-			obj[$3] = parseInt(_$1, 10);
-			return '';
-		});
+export const formatFriendlyDate = (date: Date | number | string): string => {
+	date = parseDate(date);
+	if (!date) {
 		return '';
-	});
-	obj.M--; // 月份是从0开始的，所以要减去1
-	const date = new Date(obj.y, obj.M, obj.d, obj.H, obj.m, obj.s);
-	if (obj.S !== 0) {
-		date.setMilliseconds(obj.S); // 如果设置了毫秒
 	}
-	return date;
+	const now = new Date();
+	const diff = now.getTime() - date.getTime();
+	// 1分钟以内返回 刚刚
+	if (diff < 60 * 1000) return '刚刚';
+	// 1小时以内返回 x分钟前
+	if (diff < 60 * 60 * 1000) return `${Math.floor(diff / 60000)}分钟前`;
+	const ymd = formatDate(date, 'yyyy年M月d');
+	// 当天返回 时:分
+	if (ymd === formatDate(now, 'yyyy年M月d')) return formatDate(date, 'HH:mm');
+	// 当年返回 月:日
+	if (date.getFullYear() === now.getFullYear()) return formatDate(date, 'M月d日');
+	// 否则返回 年:月:日
+	return ymd;
 };
+
 
 /**
  * 休息一段时间，单位毫秒
@@ -150,20 +192,3 @@ export const getMonthDays = (year?: number | Date, month?: number) => {
 	if (m === 1 && isLeapYear(year)) return days[m] + 1;
 	return days[m];
 };
-
-/**
- * 将一个日期格式化成友好格式，比如，1分钟以内的返回“刚刚”，
- * 当天的返回时分，当年的返回月日，否则，返回年月日
- * @param {Object} date
- */
-// formatDateToFriendly: function(date)
-// {
-// 	date = date || new Date();
-// 	date = typeof date === 'number' ? new Date(date) : date;
-// 	var now = new Date();
-// 	if((now.getTime() - date.getTime()) < 60*1000) return '刚刚'; // 1分钟以内视作“刚刚”
-// 	var temp = this.formatDate(date, 'yyyy年M月d');
-// 	if(temp == this.formatDate(now, 'yyyy年M月d')) return this.formatDate(date, 'HH:mm');
-// 	if(date.getFullYear() == now.getFullYear()) return this.formatDate(date, 'M月d日');
-// 	return temp;
-// },