+++ test/index.test.ts
@@ -1,0 +1,2 @@
+import './time.test';
+import './param.test';